+++ src/points/slice.ts
@@ -1,5 +1,5 @@
 import { PayloadAction, createSlice } from '@reduxjs/toolkit'
-import { RawClaimHistory, PointsActivityId } from 'src/points/types'
+import { RawClaimHistory, PointsActivityId, isClaimActivityId } from 'src/points/types'
 import { REHYDRATE, RehydrateAction, getRehydratePayload } from 'src/redux/persist-helper'
 
 interface GetPointsHistorySucceededAction {
@@ -77,6 +77,7 @@ const slice = createSlice({
       ...state,
       ...getRehydratePayload(action, 'points'),
       pointsConfig: { activitiesById: {} }, // always reset pointsConfig on rehydrate to ensure it's up to date
+      pointsHistory: state.pointsHistory.filter((record) => isClaimActivityId(record.activityId)), // filter in case new app version removed support for an activityId
     }))
   },
 })
